user-profile/requirements.txt
fastapi==0.115.6
pymongo==4.10.1
python-dotenv==1.0.1
uvicorn==0.32.1
azure-monitor-opentelemetry==1.6.4
fastapi-azure-auth==5.0.1
pydantic-settings==2.7.0